+++ src/features/TextToImage/components/TextToImageInputSection.vue
@@ -5,7 +5,7 @@ import {
   set,
   type Ref,
   watch,
-} from '@/library/vue';
+} from '@/library/vue/exports';
 
 import {
   VCard,
