+++ src/pages/TextToImagePage.vue
@@ -4,7 +4,7 @@ import {
   ref,
   type Ref,
   useStatefulAttemptThatEventually,
-} from '@/library/vue';
+} from '@/library/vue/exports';
 
 import {
   VBtn,
