+++ src/shims-vue.d.ts
@@ -1,7 +1,7 @@
 declare module '*.vue' {
   import type {
     DefineComponent,
-  } from '@/library/vue';
+  } from '@/library/vue/exports';
 
   const component: DefineComponent;
